+++ anxcloud/struct_disk.go
@@ -4,5 +4,5 @@ import "github.com/anexia-it/go-anxcloud/pkg/vsphere/provisioning/vm"
 
 type Disk struct {
 	*vm.Disk
-	ExactDiskSize float32
+	ExactDiskSize float64
 }
